requirements.txt
pandas
pytorch-ignite
py7zr
openpyxl